+++ packages/client/src/hooks/useTreasury.js
@@ -49,7 +49,7 @@ const doCreateTreasury = async (signerAddresses, threshold) => {
     cadence: INITIALIZE_TREASURY,
     args: (arg, t) => [
       arg(signerAddresses, t.Array(t.Address)),
-      arg(threshold, t.UInt64),
+      arg(threshold, t.Int),
     ],
     limit: CREATE_TREASURY_LIMIT,
   });
@@ -149,7 +149,7 @@ const doUpdateThreshold = async (treasuryAddr, newThreshold) => {
     cadence: UPDATE_THRESHOLD,
     args: (arg, t) => [
       arg(treasuryAddr, t.Address),
-      arg(newThreshold, t.UInt64),
+      arg(newThreshold, t.Int),
       arg(message, t.String),
       arg(keyIds, t.Array(t.UInt64)),
       arg(signatures, t.Array(t.String)),
